added style and hard coded data
